增删改查登录上传图片微信发送消息

+++ src/main/java/cn/edu/jxufe/dao/TOrderDAO.java
@@ -10,10 +10,10 @@ import java.util.List;
  * TOrderDAO继承基类
  */
 @Repository
-public interface TOrderDAO extends MyBatisBaseDao<TOrder, Integer> {
+public interface TOrderDAO extends MyBatisBaseDao<TOrder, Long> {
     public List<TOrder> findAllOrder();
     public int updateOrder(TOrder tOrder);
     public int addOrder(TOrder tOrder);
-    public int deleteOrder(int id);
-    public List<TOrder> findOrder(int t);
+    public int deleteOrder(long id);
+    public List<TOrder> findOrder(long t);
 }
